fix bug when saved password is not restored after reboot in venice appliance (#15644)
* dont set root password

* set properties from ova

* copy only if file exists

+++ tools/docker-files/vinstall/usr/local/bin/pensave-config.sh
@@ -2,5 +2,5 @@
 
 for i in /etc/passwd /etc/shadow /etc/hostname /etc/host.conf /etc/nsswitch.conf /etc/hosts
 do
-    cp -a --parents $i /config/copy-fs/
+    [[ -f $i ]] && cp -a --parents $i /config/copy-fs/
 done
